Add support for raw data type bindings
The generic binding infrastructure now allows annotating the function
argument as `str` or `bytes`.  This also adds support for `str` and
`bytes` binding for the blob functions as an example.

The binding data adapter is responsible for checking the declared
`dataType` in function.json against the declared parameter type
annotation.  All existing bindings, except blob, currently refuse to
bind if `dataType` is specified.

Issue: #66.